Fix Gemfile, add templates (partial views) for a few content block and article preview types
Rails doesn't like adding custom Bundler groups to the Gemfile (such as
:assets and :js_frameworks) so the previous (init) commit broke the entire
application. We get rid of the groups here and simply elect to comment
each block of gems.

A few template (views) have been added for some content block and article
preview types. The names of the files may be a bit ambiguous right now;
this needs to be fixed.

The idea with content blocks is that editors will have the ability to
dynamically add different blocks to their respective pages, with different
articles. For example, if they wanted to a slideshow at the top of their
pages, we'd simply render the template for that. Each (web) page will
consist of many different sections with CSS class "contentBlock."