Add preliminary authentication system on client-side (Devise on server), begin work on notifications system
Devise (specifically, devise_token_auth) has been added on the server-side to handle token-based authentication for the REST API on the staff area (client) client. In Ember, there are currently two parent routes from which all other routes (except ApplicationRoute) should extend. They are BaseRoute and UnauthenticatedRoute. An UnauthenticatedRoute route is one in which only unauthenticated (non-logged in) users may visit. Examples of these are the login, registration, and forgotten password pages. A BaseRoute route is one in which a user must be authenticated; that is, they must have a valid authentication token. We look for/validate authentication token when the application boots (page loads) and every time a client transitions into another route. If at any point a client's token becomes invalid, we save their intended destination and redirect them to the login page.

Quite a bit more work needs to be done on the client-side authentication implementation. A few to consider in the immediate future:

- Should we be looking for and/or validating a client's authentication token on every route transition? The time for transitions to complete is much larger when we validate on every transition, and this may be a problem in a production environment. There are also problems with two validation requests being sent on the "same" transition.

- The authentication token validation process is a bit quirky. I can't nail down the problem as of right now, but in some instances, the cookies storing the authentication information get deleted after a transition. This is likely due to the validation process failing, somehow.

- We need to figure out how to handle displaying error messages on the authentication/registration pages. Should we use the popup notifications system on those pages? Use a different (possibly simpler?) system?

- Decide how authentication tokens/data should be stored in production. Obviously, if a user logs in and decides to hard refresh (or open the site in a new tab/window), we don't want them to have to log in again. As of now, the authentication token, client (used by devise_token_auth to differentiate between different clients logged into the same account), and uid (the user's email) are all stored in individual cookies. Is this safe? How dangerous?

The (new) (popup) notifications system is very much a work-in-progress. In fact, much of it hasn't been implemented yet. As of now, popup notifications work by creating a popupNotificationsContainer component inside the application template, thus making it visible all throughout the app, with a visibleNotifications property, which should be a two-way binding to the controller's two-way binding to the injected notifications object's unreadNotifications (mutable) array. The popupNotificationsContainer component iterates through each object in the array and renders a popupNotification component for each one.

Popup notifications currently (should) have five properties: title, message, clickable, intent, and dismissAfter. The first two should be self-explanatory. clickable may actually be a bit redundant here, as a popup notification is only clickable if it has an intent. intent should be an anonymous function that is executed when the popup notification is clicked (hence when clickable is quite redundant). dismissAfter is the time (in seconds) for which the notification should stay visible. After that, it is dismissed. We should change the unreadNotifications property on the notifications object to something a bit more close its actual purpose, such as visiblePopupNotifications. Popup notifications that aren't manually dismissed by the user should remain an unreadNotification. Thus, a popup notification can be both unread yet also already "popped up."

Some other small changes also made it into this commit, not really worthy of talking about. This commit should really have been split up into multiple for the sake of good practices.